feat: adds ecosystem to projectList trpc API response
- also moves type MyForestProject into common types for My Forest v2
